+++ app/models/option_models.py
@@ -1,15 +1,23 @@
+"""
+The module defines the SQLAlchemy model for the option entity, which is
+used to store and manage configuration values for both the application
+and its extensions within the database.
+"""
+
 from sqlalchemy import Column, BigInteger, Integer, String, ForeignKey
 from sqlalchemy.orm import relationship
-from app.config import get_config
 from app.database import Base
 from time import time
 
-cfg = get_config()
-
 
 class Option(Base):
+    """
+    SQLAlchemy model for an option entity. The model is used to store
+    and manage configuration values for both the application and its
+    extensions within the database.
+    """
     __tablename__ = "options"
-    _cacheable = True
+    _cacheable = False
 
     id = Column(BigInteger, primary_key=True)
     created_date = Column(Integer, index=True, default=lambda: int(time()))
@@ -20,20 +28,26 @@ class Option(Base):
     option_value = Column(String(512), nullable=False, index=True)
 
     option_user = relationship(
-        "User", back_populates="user_options", lazy="joined")
+        "User", back_populates="user_options", lazy="noload")
 
     def __init__(self, user_id: int, option_key: str, option_value: str):
+        """
+        Initializes an option instance with the specified user ID,
+        option key, and option value.
+        """
         self.user_id = user_id
         self.option_key = option_key
         self.option_value = option_value
 
     def to_dict(self):
+        """
+        Converts the option instance to a dictionary representation,
+        including all necessary attributes.
+        """
         return {
             "id": self.id,
             "created_date": self.created_date,
             "updated_date": self.updated_date,
-            "user_id": self.user_id,
-            "option_key": self.option_value,
+            "option_key": self.option_key,
             "option_value": self.option_value,
-            "comment_user": self.comment_user.to_dict(),
         }